fix/feat: simplifies code by adding date function to our lib folder and importing them, also removes the random add to tsconfig.json
