+++ src/app/hours/page.tsx
@@ -15,63 +15,8 @@ import React, { useEffect, useState } from 'react';
 import style from '@styles/admin/users.module.css';
 import Link from 'next/link';
 import { useUser } from '@clerk/clerk-react';
-import { IEvent } from '@database/eventSchema';
-
-//convert date into format Dayofweek, Month
-const formatDate = (date: Date) => {
-  if (!(date instanceof Date)) {
-    date = new Date(date); // Convert to Date object if not already
-  }
-
-  const options: Intl.DateTimeFormatOptions = {
-    month: 'numeric',
-    day: 'numeric',
-    year: 'numeric',
-  };
-  return date.toLocaleDateString('en-US', options);
-};
-
-// convert date into xx:xx XM - xx:xx XM
-const formatDuration = (start: Date, end: Date) => {
-  if (!(start instanceof Date)) {
-    start = new Date(start); // Convert to Date object if not already
-  }
-
-  if (!(end instanceof Date)) {
-    end = new Date(end); // Convert to Date object if not already
-  }
-
-  // Calculate the time difference in milliseconds
-  const timeDiff = end.getTime() - start.getTime();
-
-  // Convert milliseconds to minutes
-  const minutes = Math.floor(timeDiff / (1000 * 60));
-
-  // Calculate remaining hours and minutes
-  const hours = Math.floor(minutes / 60);
-  const remainingMinutes = minutes % 60;
-
-  return `${hours}h ${remainingMinutes} min`;
-};
-
-// convert date into xx:xx XM - xx:xx XM
-const getDuration = (start: Date, end: Date) => {
-  if (!(start instanceof Date)) {
-    start = new Date(start); // Convert to Date object if not already
-  }
-
-  if (!(end instanceof Date)) {
-    end = new Date(end); // Convert to Date object if not already
-  }
-
-  // Calculate the time difference in milliseconds
-  const timeDiff = end.getTime() - start.getTime();
-
-  // Convert milliseconds to minutes
-  const minutes = Math.floor(timeDiff / (1000 * 60));
-
-  return minutes;
-};
+import { IEvent } from '../../database/eventSchema';
+import { formatDate, formatDuration, getDuration } from '../lib/dates';
 
 const AttendedEvents = () => {
   //states
